Feeling parameter

+++ Activities.Strava/Endpoints/Models/Gear.cs
@@ -4,19 +4,14 @@ namespace Activities.Strava.Endpoints.Models
 {
     public class Gear
     {
-        [JsonProperty("id")]
-        public string Id { get; set; }
+        [JsonProperty("id")] public string Id { get; set; }
 
-        [JsonProperty("primary")]
-        public bool Primary { get; set; }
+        [JsonProperty("primary")] public bool Primary { get; set; }
 
-        [JsonProperty("name")]
-        public string Name { get; set; }
+        [JsonProperty("name")] public string Name { get; set; }
 
-        [JsonProperty("resource_state")]
-        public int ResourceState { get; set; }
+        [JsonProperty("resource_state")] public int ResourceState { get; set; }
 
-        [JsonProperty("distance")]
-        public int Distance { get; set; }
+        [JsonProperty("distance")] public double Distance { get; set; }
     }
 }
